requirements.txt
Flask
Flask-JWT-Extended
pymongo
Flask-CORS
assemblyai
google-generativeai
gTTS
gunicorn
python-dotenv